update file names for experience and education
