+++ services/frontend/src/components/educationCard/EducationCardView.jsx
@@ -1,15 +1,11 @@
 import { FormattedMessage } from "react-intl";
-import {
-  ContainerOutlined,
-  FileDoneOutlined,
-  LinkOutlined,
-} from "@ant-design/icons";
+import { BankOutlined, LinkOutlined } from "@ant-design/icons";
 import { Avatar, Empty, List, Tag } from "antd";
 import PropTypes from "prop-types";
 
-import "./ExperienceView.less";
+import "./EducationCardView.less";
 
-const ExperienceView = ({ experienceInfo }) => {
+const EducationCardView = ({ educationInfo }) => {
   /**
    * Generate styled description text
    * @param {string} text - text to display as description
@@ -37,16 +33,9 @@ const ExperienceView = ({ experienceInfo }) => {
     SupportingLinks &&
     SupportingLinks.length > 0 && (
       <>
-        <div className="d-block">
-          <LinkOutlined
-            aria-hidden="true"
-            className="mr-1 d-inline"
-            style={{ color: "#3CBAB3" }}
-          />
-          <h5 className="mt-1 d-inline">
-            <FormattedMessage id="attachment.links.employment" />
-          </h5>
-        </div>
+        <h5 className="visually-hidden">
+          <FormattedMessage id="attachment.links.education" />
+        </h5>
         {SupportingLinks.map((link) => (
           <a href={link.url} rel="noopener noreferrer" target="_blank">
             <Tag key={link.id} color="#727272" style={{ cursor: "pointer" }}>
@@ -61,48 +50,20 @@ const ExperienceView = ({ experienceInfo }) => {
       </>
     );
 
-  /**
-   * Generate the list of projects
-   * @param {Array.<string>} projects - list of project names
-   * @returns {React.ReactElement} - React Element
-   */
-  const generateProjectsList = (projects) =>
-    projects &&
-    projects.length > 0 && (
-      <div className="mb-1">
-        <div className="d-block">
-          <FileDoneOutlined
-            aria-hidden="true"
-            className="mr-1 d-inline"
-            style={{ color: "#3CBAB3" }}
-          />
-          <h5 className="mt-1 d-inline">
-            <FormattedMessage id="projects" />
-          </h5>
-        </div>
-        {projects.map((i) => (
-          <Tag key={i} color="#727272">
-            <FileDoneOutlined aria-hidden="true" className="mr-1" />
-            {i}
-          </Tag>
-        ))}
-      </div>
-    );
-
-  if (experienceInfo && experienceInfo.length > 0) {
+  if (educationInfo && educationInfo.length > 0) {
     return (
       <List
-        dataSource={experienceInfo}
+        dataSource={educationInfo}
         itemLayout="vertical"
-        renderItem={(item) => (
+        renderItem={(educationItem) => (
           <List.Item
-            className="experience-item-list"
+            className="education-item-list"
             extra={
               <>
                 <h5 className="visually-hidden">
                   <FormattedMessage id="duration" />
                 </h5>
-                {item.duration}
+                {educationItem.duration}
               </>
             }
           >
@@ -110,22 +71,20 @@ const ExperienceView = ({ experienceInfo }) => {
               avatar={
                 <Avatar
                   aria-hidden="true"
-                  className="experience-avatar"
-                  icon={<ContainerOutlined aria-hidden="true" />}
+                  className="education-avatar"
+                  icon={<BankOutlined aria-hidden="true" />}
                   shape="square"
                   size="large"
                 />
               }
-              description={item.organization}
-              title={item.jobTitle}
+              description={educationItem.school}
+              title={educationItem.diploma}
             />
+            {educationItem.description &&
+              generateDescriptionBody(educationItem.description)}
 
-            {item.description && generateDescriptionBody(item.description)}
-
-            {item.projects && generateProjectsList(item.projects)}
-
-            {item.attachmentLinks &&
-              generateSupportingLinks(item.attachmentLinks)}
+            {educationItem.attachmentLinks &&
+              generateSupportingLinks(educationItem.attachmentLinks)}
           </List.Item>
         )}
       />
@@ -133,14 +92,14 @@ const ExperienceView = ({ experienceInfo }) => {
   }
   return (
     <Empty
-      description={<FormattedMessage id="experience.empty" />}
+      description={<FormattedMessage id="education.empty" />}
       image={Empty.PRESENTED_IMAGE_SIMPLE}
     />
   );
 };
 
-ExperienceView.propTypes = {
-  experienceInfo: PropTypes.arrayOf(
+EducationCardView.propTypes = {
+  educationInfo: PropTypes.arrayOf(
     PropTypes.shape({
       attachmentLinks: PropTypes.arrayOf(
         PropTypes.shape({
@@ -150,12 +109,11 @@ ExperienceView.propTypes = {
         })
       ),
       description: PropTypes.string,
+      diploma: PropTypes.string,
       duration: PropTypes.string,
-      jobTitle: PropTypes.string,
-      organization: PropTypes.string,
-      projects: PropTypes.arrayOf(PropTypes.string),
+      school: PropTypes.string,
     })
   ).isRequired,
 };
 
-export default ExperienceView;
+export default EducationCardView;
