+++ services/frontend/src/components/educationCard/EducationCardView.less
