+++ services/frontend/src/components/experienceCard/ExperienceCardView.jsx
@@ -1,11 +1,15 @@
 import { FormattedMessage } from "react-intl";
-import { BankOutlined, LinkOutlined } from "@ant-design/icons";
+import {
+  ContainerOutlined,
+  FileDoneOutlined,
+  LinkOutlined,
+} from "@ant-design/icons";
 import { Avatar, Empty, List, Tag } from "antd";
 import PropTypes from "prop-types";
 
-import "./EducationView.less";
+import "./ExperienceCardView.less";
 
-const EducationView = ({ educationInfo }) => {
+const ExperienceCardView = ({ experienceInfo }) => {
   /**
    * Generate styled description text
    * @param {string} text - text to display as description
@@ -33,9 +37,16 @@ const EducationView = ({ educationInfo }) => {
     SupportingLinks &&
     SupportingLinks.length > 0 && (
       <>
-        <h5 className="visually-hidden">
-          <FormattedMessage id="attachment.links.education" />
-        </h5>
+        <div className="d-block">
+          <LinkOutlined
+            aria-hidden="true"
+            className="mr-1 d-inline"
+            style={{ color: "#3CBAB3" }}
+          />
+          <h5 className="mt-1 d-inline">
+            <FormattedMessage id="attachment.links.employment" />
+          </h5>
+        </div>
         {SupportingLinks.map((link) => (
           <a href={link.url} rel="noopener noreferrer" target="_blank">
             <Tag key={link.id} color="#727272" style={{ cursor: "pointer" }}>
@@ -50,20 +61,48 @@ const EducationView = ({ educationInfo }) => {
       </>
     );
 
-  if (educationInfo && educationInfo.length > 0) {
+  /**
+   * Generate the list of projects
+   * @param {Array.<string>} projects - list of project names
+   * @returns {React.ReactElement} - React Element
+   */
+  const generateProjectsList = (projects) =>
+    projects &&
+    projects.length > 0 && (
+      <div className="mb-1">
+        <div className="d-block">
+          <FileDoneOutlined
+            aria-hidden="true"
+            className="mr-1 d-inline"
+            style={{ color: "#3CBAB3" }}
+          />
+          <h5 className="mt-1 d-inline">
+            <FormattedMessage id="projects" />
+          </h5>
+        </div>
+        {projects.map((i) => (
+          <Tag key={i} color="#727272">
+            <FileDoneOutlined aria-hidden="true" className="mr-1" />
+            {i}
+          </Tag>
+        ))}
+      </div>
+    );
+
+  if (experienceInfo && experienceInfo.length > 0) {
     return (
       <List
-        dataSource={educationInfo}
+        dataSource={experienceInfo}
         itemLayout="vertical"
-        renderItem={(educationItem) => (
+        renderItem={(item) => (
           <List.Item
-            className="education-item-list"
+            className="experience-item-list"
             extra={
               <>
                 <h5 className="visually-hidden">
                   <FormattedMessage id="duration" />
                 </h5>
-                {educationItem.duration}
+                {item.duration}
               </>
             }
           >
@@ -71,20 +110,22 @@ const EducationView = ({ educationInfo }) => {
               avatar={
                 <Avatar
                   aria-hidden="true"
-                  className="education-avatar"
-                  icon={<BankOutlined aria-hidden="true" />}
+                  className="experience-avatar"
+                  icon={<ContainerOutlined aria-hidden="true" />}
                   shape="square"
                   size="large"
                 />
               }
-              description={educationItem.school}
-              title={educationItem.diploma}
+              description={item.organization}
+              title={item.jobTitle}
             />
-            {educationItem.description &&
-              generateDescriptionBody(educationItem.description)}
 
-            {educationItem.attachmentLinks &&
-              generateSupportingLinks(educationItem.attachmentLinks)}
+            {item.description && generateDescriptionBody(item.description)}
+
+            {item.projects && generateProjectsList(item.projects)}
+
+            {item.attachmentLinks &&
+              generateSupportingLinks(item.attachmentLinks)}
           </List.Item>
         )}
       />
@@ -92,14 +133,14 @@ const EducationView = ({ educationInfo }) => {
   }
   return (
     <Empty
-      description={<FormattedMessage id="education.empty" />}
+      description={<FormattedMessage id="experience.empty" />}
       image={Empty.PRESENTED_IMAGE_SIMPLE}
     />
   );
 };
 
-EducationView.propTypes = {
-  educationInfo: PropTypes.arrayOf(
+ExperienceCardView.propTypes = {
+  experienceInfo: PropTypes.arrayOf(
     PropTypes.shape({
       attachmentLinks: PropTypes.arrayOf(
         PropTypes.shape({
@@ -109,11 +150,12 @@ EducationView.propTypes = {
         })
       ),
       description: PropTypes.string,
-      diploma: PropTypes.string,
       duration: PropTypes.string,
-      school: PropTypes.string,
+      jobTitle: PropTypes.string,
+      organization: PropTypes.string,
+      projects: PropTypes.arrayOf(PropTypes.string),
     })
   ).isRequired,
 };
 
-export default EducationView;
+export default ExperienceCardView;
