+++ services/frontend/src/components/experienceCard/ExperienceCardView.less
